+++ src/ui/dialogs/configdialog.h
@@ -68,7 +68,7 @@ private:
     QTabWidget* m_mainwidget;
     QDialogButtonBox* m_buttons;
     QComboBox* m_charttheme;
-    QCheckBox *m_animated_charts, *m_auto_confidence, *m_tooltips, *m_ask_on_exit, *m_save_on_exit, *m_series_confidence, *m_RemoveGrid, *m_EmphAxis, *m_auto_thermo_dialog;
+    QCheckBox *m_animated_charts, *m_auto_confidence, *m_tooltips, *m_ask_on_exit, *m_save_on_exit, *m_series_confidence, *m_RemoveGrid, *m_EmphAxis, *m_auto_thermo_dialog, *m_thermogram_guideline;
 
     void setUi();
     void createGeneralTab();