+++ src/ui/widgets/thermogramwidget.cpp
@@ -68,6 +68,15 @@ ThermogramWidget::ThermogramWidget(QWidget* parent)
 
 void ThermogramWidget::setUi()
 {
+    m_guide_label = new QLabel;
+    if (qApp->instance()->property("thermogram_guidelines").toBool()) {
+        m_guide_label->setFixedHeight(70);
+        m_guide_label->setWordWrap(true);
+        m_guide_label->setMaximumWidth(1100);
+        m_guide_label->setToolTip(tr("Hey, I am like Clippit (german: Karl Klammer). If you do not want to see me, just disable me in the general config (=> [General Settings] => [Show guidelines in Thermogram Dialog])!"));
+    } else
+        m_guide_label->setFixedHeight(30);
+
     m_thermogram = new ChartView;
     m_thermogram->setAutoScaleStrategy(AutoScaleStrategy::QtNiceNumbers);
 
@@ -85,6 +94,7 @@ void ThermogramWidget::setUi()
         m_peak_start_line->hide();
         m_peak_end_line->hide();
         m_peak_edit_mode = false;
+        ResetGuideLabel();
     });
 
     connect(m_thermogram, &ChartView::RightKey, this, [this]() {
@@ -106,11 +116,14 @@ void ThermogramWidget::setUi()
 
     m_table = new QTableWidget;
     connect(m_table, &QTableWidget::doubleClicked, this, QOverload<const QModelIndex&>::of(&ThermogramWidget::PeakDoubleClicked));
+    m_table->setToolTip(tr("This table hold all peaks, consisting of the integral, start time and end time. Double click on a peak will toggle the <i>Peak Integration mode</i>. Double click of within the thermogram selects the appropriate peak in this table (if any peaks already exists"));
 
     m_peak_rule_list = new QTableWidget;
     m_peak_rule_list->setRowCount(1);
     m_peak_rule_list->setColumnCount(2);
     m_peak_rule_list->setContextMenuPolicy(Qt::ActionsContextMenu);
+    m_peak_rule_list->setToolTip(tr("This table holds the rules (start time and peak duration) to generate the peak list. One rule will be repeated until the next start time (in the next peak rule) is due. This rule will then be applied. Incomplete peaks will be ignored. If there is the last peak incomplete, add the starting point as rule and reduce the peak duration."));
+
     QAction* action = new QAction("Add Rule");
     action->setIcon(Icon("list-add"));
     connect(action, &QAction::triggered, m_peak_rule_list, [this]() {
@@ -125,6 +138,7 @@ void ThermogramWidget::setUi()
     m_peak_rule_list->addAction(action);
 
     action = new QAction("Remove (last) Rule");
+    action->setIcon(Icon("trash-empty"));
     connect(action, &QAction::triggered, m_peak_rule_list, [this]() {
         if (m_peak_rule_list->rowCount() > 1) {
             //int rows = m_peak_rule_list->rowCount() -2;
@@ -136,8 +150,8 @@ void ThermogramWidget::setUi()
 
     connect(m_peak_rule_list, &QTableWidget::doubleClicked, this, QOverload<const QModelIndex&>::of(&ThermogramWidget::PeakRuleDoubleClicked));
 
-    QStringList header = QStringList() << "start"
-                                       << "interval";
+    QStringList header = QStringList() << "Start Time\n[s]"
+                                       << "Peak Duration\n[s]";
     m_peak_rule_list->setHorizontalHeaderLabels(header);
 
     QTabWidget* peaks_tab = new QTabWidget;
@@ -160,6 +174,7 @@ void ThermogramWidget::setUi()
 
     QGridLayout* baselayout = new QGridLayout;
 
+    /*
     m_baseline_type = new QComboBox;
     QStringList options = QStringList() << tr("none") << tr("offset") << tr("constant") << tr("polynomial");
     m_baseline_type->addItems(options);
@@ -183,6 +198,7 @@ void ThermogramWidget::setUi()
     baselayout->addWidget(m_poly_slow, 0, 4);
 
     m_limits = new QCheckBox(tr("Show Fit limit"));
+    */
     /*
     baselayout->addWidget(m_limits, 0, 5);
     connect(m_limits, &QCheckBox::stateChanged, this, [this](int state) {
@@ -191,6 +207,7 @@ void ThermogramWidget::setUi()
         UpdateLimits();
     });
     */
+    /*
     m_smooth = new QCheckBox(tr("Smooth"));
     connect(m_smooth, &QCheckBox::stateChanged, this, &ThermogramWidget::UpdatePlot);
     connect(m_smooth, &QCheckBox::stateChanged, this, &ThermogramWidget::Update);
@@ -215,25 +232,27 @@ void ThermogramWidget::setUi()
     baselayout->addWidget(m_constant, 1, 1);
     baselayout->addWidget(new QLabel(tr("Coefficents:")), 1, 2);
     baselayout->addWidget(m_coeffs, 1, 3);
+    */
     /*
     baselayout->addWidget(m_smooth, 1, 4);
     baselayout->addWidget(m_filter, 1, 5);
     */
-
+    /*
     m_stdev = new QLineEdit;
     connect(m_stdev, &QLineEdit::textChanged, this, &ThermogramWidget::UpdateLimits);
 
     m_mult = new QLineEdit("1");
     connect(m_mult, &QLineEdit::textChanged, this, &ThermogramWidget::UpdateLimits);
-
+    */
     /*
     baselayout->addWidget(new QLabel(tr("Stddev")), 2, 0);
     baselayout->addWidget(m_stdev, 2, 1);
     baselayout->addWidget(new QLabel(tr("Multiplier")), 2, 2);
     baselayout->addWidget(m_mult, 2, 3);
     */
+    /*
     m_fit_button = new QPushButton(tr("Refit Baseline"));
-    // baselayout->addWidget(m_fit_button, 2, 4);
+    baselayout->addWidget(m_fit_button, 2, 4);
     connect(m_fit_button, &QPushButton::clicked, this, &ThermogramWidget::FitBaseLine);
 
     m_baseline_polynom = new QPlainTextEdit;
@@ -258,43 +277,70 @@ void ThermogramWidget::setUi()
         m_constant->setDisabled(!m_full_spec->isChecked());
         FitBaseLine();
     });
-
+    */
+    /*
     QHBoxLayout* base_input = new QHBoxLayout;
     base_input->addWidget(m_full_spec);
     base_input->addWidget(m_peak_wise);
     base_input->addWidget(m_fit_button);
     m_peak_wise->setChecked(true);
     m_full_spec->setChecked(false);
+    */
 
     m_const_offset = new QDoubleSpinBox;
     m_const_offset->setMinimum(-1e5);
     m_const_offset->setMaximum(1e5);
     m_const_offset->setValue(0);
     m_const_offset->setDecimals(7);
+
     connect(m_const_offset, &QDoubleSpinBox::editingFinished, this, &ThermogramWidget::Update);
+    connect(m_const_offset, &QDoubleSpinBox::editingFinished, this, &ThermogramWidget::ResetGuideLabel);
+
+    connect(m_const_offset, qOverload<double>(&QDoubleSpinBox::valueChanged), this, [this]() {
+        setGuideText("Press [Enter] to apply the current value which will be removed from your integrals. Please keep in mind, that the dimension of the value should be the same as within this tabulator - thus please use the raw unit observed from your calorimeter and not converted units, for example from cal into J.");
+    });
 
     m_peaks_start = new QDoubleSpinBox;
     m_peaks_start->setMinimum(0);
     m_peaks_start->setMaximum(1e8);
-    //connect(m_peaks_start, &QDoubleSpinBox::editingFinished, this, &ThermogramWidget::Divide2Peaks);
 
-    m_get_peaks_start = new QPushButton(tr("Select"));
+    m_get_peaks_start = new QPushButton(tr("Click to Select"));
+    m_get_peaks_start->setIcon(Icon("edit-select"));
+    connect(m_thermogram, &ChartView::PointDoubleClicked, this, &ThermogramWidget::PointDoubleClicked);
+
     connect(m_get_peaks_start, &QPushButton::clicked, this, [this]() {
-        m_get_time_from_thermogram = 1;
-        QApplication::setOverrideCursor(QCursor(Qt::CrossCursor));
-        connect(m_thermogram, &ChartView::PointDoubleClicked, this, &ThermogramWidget::PointDoubleClicked);
+        if (!m_get_time_from_thermogram) {
+            m_get_time_from_thermogram = 1;
+            QApplication::setOverrideCursor(QCursor(Qt::CrossCursor));
+            m_get_peaks_start->setText("Click to Undo");
+            setGuideText(QString("You are now in <i>Thermogram start selection mode</i>. <b>Double click</b> with the <b>left</b> mouse in the chart to define, where the peak starts. Click on the <em>Click to Undo Button</em> to decline the selection. Zooming is still possible via <b>single left click</b>."));
+        } else if (m_get_time_from_thermogram == 1) {
+            m_get_time_from_thermogram = 0;
+            m_get_peaks_start->setText("Click to Select");
+            ResetGuideLabel();
+            QApplication::restoreOverrideCursor();
+        }
+
     });
 
     m_peaks_end = new QDoubleSpinBox;
     m_peaks_end->setMinimum(0);
     m_peaks_end->setMaximum(1e8);
-    connect(m_peaks_end, &QDoubleSpinBox::editingFinished, this, &ThermogramWidget::Divide2Peaks);
 
-    m_get_peaks_end = new QPushButton(tr("Select"));
+    m_get_peaks_end = new QPushButton(tr("Click to Select"));
+    m_get_peaks_end->setIcon(Icon("edit-select"));
     connect(m_get_peaks_end, &QPushButton::clicked, this, [this]() {
-        m_get_time_from_thermogram = 2;
-        QApplication::setOverrideCursor(QCursor(Qt::CrossCursor));
-        connect(m_thermogram, &ChartView::PointDoubleClicked, this, &ThermogramWidget::PointDoubleClicked);
+        if (!m_get_time_from_thermogram) {
+            m_get_time_from_thermogram = 2;
+            QApplication::setOverrideCursor(QCursor(Qt::CrossCursor));
+            m_get_peaks_end->setText("Click to Undo");
+            setGuideText(QString("You are now in <i>Thermogram end selection mode</i>. <b>Double click</b> with the <b>left</b> mouse in the chart to define, where the peak starts. Click on the <em>Click to Undo Button</em> to decline the selection. Zooming is still possible via <b>single left click</b>."));
+        } else if (m_get_time_from_thermogram == 2) {
+            m_get_time_from_thermogram = 0;
+            m_get_peaks_end->setText("Click to Select");
+            ResetGuideLabel();
+            QApplication::restoreOverrideCursor();
+        }
     });
 
     m_peaks_time = new QDoubleSpinBox;
@@ -312,6 +358,8 @@ void ThermogramWidget::setUi()
         CalibrateSystem();
         ApplyCalibration();
     });
+    m_calibration_start->setMaximumWidth(100);
+    m_calibration_start->setToolTip(tr("Set the duration in seconds of the calibration peak. The calibration peak is considered to be last peak in the thermogram. Leave zero, if the calorimeter does not use calibration peaks"));
 
     m_calibration_heat = new QDoubleSpinBox;
     m_calibration_heat->setMinimum(-1e8);
@@ -324,15 +372,22 @@ void ThermogramWidget::setUi()
         CalibrateSystem();
         ApplyCalibration();
     });
+    m_calibration_heat->setMaximumWidth(100);
+    m_calibration_heat->setToolTip(tr("Set the heat in J, that were used to for the calibration. Leave zero if no calibration peak was used. If the value is not zero, the scaling factor may be redundant. Please keep that in mind."));
 
-    m_calibration_label = new QLabel(tr("Calibration (0)"));
+    m_calibration_label = new QLabel(tr("<h4>Calibration (0)</h4>"));
 
     m_integration_range = new QComboBox;
     m_integration_range->addItems(m_Peak_Cut_Options);
+    m_integration_range->setMaximumWidth(100);
 
     m_peak_apply = new QPushButton(tr("Apply and Update"));
+    m_peak_apply->setIcon(Icon("dialog-ok-apply"));
     connect(m_peak_apply, &QPushButton::clicked, this, &ThermogramWidget::UpdatePeaks);
+    m_peak_apply->setMaximumWidth(150);
+    m_peak_apply->setToolTip(tr("Click generate peaks, fit baseline and integrate the peaks"));
 
+    /*
     m_get_peaks_range = new QPushButton(tr("Select Peak Range"));
     connect(m_get_peaks_range, &QPushButton::clicked, this, [this]() {
         m_thermogram->setSelectStrategy(SelectStrategy::S_Horizontal);
@@ -347,50 +402,70 @@ void ThermogramWidget::setUi()
 
     m_auto_pick = new QPushButton(tr("Auto Picking"));
     connect(m_auto_pick, &QPushButton::clicked, this, &ThermogramWidget::PickPeaks);
+    */
 
     m_integration_range_threshold = new QDoubleSpinBox;
     m_integration_range_threshold->setMinimum(0);
     m_integration_range_threshold->setMaximum(1e5);
     m_integration_range_threshold->setDecimals(10);
+    m_integration_range_threshold->setValue(0);
+    m_integration_range_threshold->setMaximumWidth(150);
+    m_integration_range_threshold->setToolTip(tr("Define a threshold for Peak Range Reduction"));
 
     m_iterations = new QSpinBox;
     m_iterations->setMinimum(1);
     m_iterations->setMaximum(1e3);
     m_iterations->setValue(15);
-    //connect(m_integration_range_threshold, &QDoubleSpinBox::editingFinished, this, &ThermogramWidget::CutAllLimits);
+    m_iterations->setToolTip(tr("Define the maximum number of iteration for the Peak Range Reduction Cycle. In some case, this value should be 1."));
 
     QHBoxLayout* peak_layout = new QHBoxLayout;
+    QVBoxLayout* vlayout = new QVBoxLayout;
+
+    vlayout->addWidget(new QLabel(tr("Remove constant")));
+    vlayout->addWidget(m_const_offset);
+    peak_layout->addLayout(vlayout);
+
     /*
-    peak_layout->addWidget(new QLabel(tr("Remove constant")));
-    peak_layout->addWidget(m_const_offset);
     peak_layout->addWidget(m_peak_sensitivity);
     peak_layout->addWidget(m_auto_pick);
     peak_layout->addWidget(new QLabel(tr("Peak Count")));
     peak_layout->addWidget(m_peak_count);
     */
 
-    QVBoxLayout* vlayout = new QVBoxLayout;
-    vlayout->addWidget(new QLabel(tr("Thermogram starts (s)")));
+    QHBoxLayout* triplett = new QHBoxLayout;
+
+    vlayout = new QVBoxLayout;
+    triplett->addWidget(new QLabel(tr("Thermogram\nstart time [s]")));
     vlayout->addWidget(m_get_peaks_start);
     vlayout->addWidget(m_peaks_start);
-    peak_layout->addLayout(vlayout);
+    triplett->addLayout(vlayout);
+    peak_layout->addLayout(triplett);
 
+    triplett = new QHBoxLayout;
+    triplett->addWidget(new QLabel(tr("Thermogram\nend time [s]")));
     vlayout = new QVBoxLayout;
-    vlayout->addWidget(new QLabel(tr("Thermogram ends (s)")));
     vlayout->addWidget(m_get_peaks_end);
     vlayout->addWidget(m_peaks_end);
-    peak_layout->addLayout(vlayout);
+    triplett->addLayout(vlayout);
+    peak_layout->addLayout(triplett);
 
-    vlayout = new QVBoxLayout;
-    vlayout->addWidget(new QLabel(tr("Time between injects (s)")));
-    vlayout->addWidget(m_peaks_time);
-    vlayout->addWidget(m_get_peaks_range);
-    peak_layout->addLayout(vlayout);
+    triplett = new QHBoxLayout;
+    triplett->addWidget(new QLabel(tr("Time between\ninjections [s]")));
+    triplett->addWidget(m_peaks_time);
+    peak_layout->addLayout(triplett);
 
     vlayout = new QVBoxLayout;
     vlayout->addWidget(m_calibration_label);
-    vlayout->addWidget(m_calibration_start);
-    vlayout->addWidget(m_calibration_heat);
+
+    triplett = new QHBoxLayout;
+    triplett->addWidget(new QLabel("Time start [s]"));
+    triplett->addWidget(m_calibration_start);
+    vlayout->addLayout(triplett);
+
+    triplett = new QHBoxLayout;
+    triplett->addWidget(new QLabel("Heat [raw]"));
+    triplett->addWidget(m_calibration_heat);
+    vlayout->addLayout(triplett);
     peak_layout->addLayout(vlayout);
 
     vlayout = new QVBoxLayout;
@@ -400,24 +475,36 @@ void ThermogramWidget::setUi()
     iterlayout->addWidget(new QLabel(tr("Iterations")));
     iterlayout->addWidget(m_iterations);
     vlayout->addLayout(iterlayout);
-    vlayout->addWidget(m_integration_range);
+
+    iterlayout = new QHBoxLayout;
+    iterlayout->addWidget(new QLabel(tr("Method")));
+    iterlayout->addWidget(m_integration_range);
+    vlayout->addLayout(iterlayout);
+
     vlayout->addWidget(m_peak_apply);
     peak_layout->addLayout(vlayout);
 
     QVBoxLayout* layout = new QVBoxLayout;
     layout->addLayout(peak_layout);
+    layout->addWidget(m_guide_label);
     layout->addWidget(chart);
+
+#pragma message("this disabled, since it has not proven as being useful for now")
+    /*
     layout->addLayout(base_input);
     layout->addWidget(baseline);
-
+    */
     setLayout(layout);
 
+    /*
+    
     m_fit_type->setDisabled(true);
     m_coeffs->setDisabled(false);
     m_constant->setDisabled(true);
 
     m_stdev->setEnabled(false);
     m_mult->setEnabled(false);
+    */
 
     QSettings settings;
     settings.beginGroup("thermogram");
@@ -434,6 +521,10 @@ void ThermogramWidget::setUi()
     connect(m_integration_range, &QComboBox::currentTextChanged, this, &ThermogramWidget::CutAllLimits);
     connect(m_peaks_start, qOverload<double>(&QDoubleSpinBox::valueChanged), this, [this](double value) {
         m_peak_rule_list->item(m_current_peaks_rule, 0)->setData(Qt::DisplayRole, value);
+    });
+
+    connect(m_peaks_time, qOverload<double>(&QDoubleSpinBox::valueChanged), this, [this](double value) {
+        m_peak_rule_list->item(m_current_peaks_rule, 1)->setData(Qt::DisplayRole, value);
     });
 }
 
@@ -467,7 +558,7 @@ void ThermogramWidget::setThermogram(PeakPick::spectrum* spec, qreal offset)
     std::vector<PeakPick::Peak> peaks = PeakPick::PickPeaks(&sign, 0, qPow(2, 1));
     std::vector<PeakPick::Peak> max_peak = PeakPick::PickPeaks(&m_spec, 0, qPow(2, 1));
 
-    m_peak_count->setValue(peaks.size() + max_peak.size());
+    // m_peak_count->setValue(peaks.size() + max_peak.size());
 
     m_peaks_start->setMaximum(m_spec.XMax());
     m_peaks_end->setMaximum(m_spec.XMax() - m_calibration_start->value() * m_spec.Step());
@@ -476,8 +567,8 @@ void ThermogramWidget::setThermogram(PeakPick::spectrum* spec, qreal offset)
     m_peaks_end->setValue(m_spec.XMax() - m_calibration_start->value() * m_spec.Step());
 
     UpdatePlot();
-    m_stdev->setText(QString::number(spec->StdDev()));
-    m_constant->setText(QString::number(spec->Mean()));
+    // m_stdev->setText(QString::number(spec->StdDev()));
+    // m_constant->setText(QString::number(spec->Mean()));
 }
 
 void ThermogramWidget::setPeakList(const std::vector<PeakPick::Peak>& peak_list)
@@ -488,7 +579,7 @@ void ThermogramWidget::setPeakList(const std::vector<PeakPick::Peak>& peak_list)
         m_peaks_time->setValue(m_peak_list[0].end * m_spec.Step() - m_peak_list[0].start * m_spec.Step() + m_spec.Step());
         m_peaks_end->setValue(m_peak_list[m_peak_list.size() - 1].end * m_spec.Step());
     }
-    m_peak_count->setValue(m_peak_list.size());
+    // m_peak_count->setValue(m_peak_list.size());
     UpdatePeaks();
     // FitBaseLine();
 }
@@ -536,7 +627,7 @@ void ThermogramWidget::UpdateTable()
                                        << "peak end\n[s]";
     m_table->setHorizontalHeaderLabels(header);
     m_table->resizeColumnsToContents();
-
+    /*
     if (m_baseline.baselines.size() == 1) {
         QString string("<html><p>f(x) = ");
         for (unsigned int i = 0; i < m_baseline.baselines[0].size(); ++i) {
@@ -555,6 +646,7 @@ void ThermogramWidget::UpdateTable()
         m_baseline_polynom->appendHtml(string);
     } else
         m_baseline_polynom->clear();
+    */
 }
 
 void ThermogramWidget::UpdatePlot()
@@ -588,6 +680,7 @@ void ThermogramWidget::UpdateLimits()
     m_thermogram->addSeries(m_lower);
 }
 
+/*
 void ThermogramWidget::PickPeaks()
 {
     if (!m_spectrum)
@@ -613,12 +706,12 @@ void ThermogramWidget::PickPeaks()
     peaks.insert(peaks.end(), max_peak.begin(), max_peak.end());
     setPeakList(peaks);
 }
-
+*/
 void ThermogramWidget::fromSpectrum(const PeakPick::spectrum* original, LineSeries* series)
 {
     PeakPick::spectrum* spectrum = new PeakPick::spectrum(original);
-    if (m_smooth->isChecked())
-        SmoothFunction(spectrum, m_filter->value());
+    //    if (m_smooth->isChecked())
+    //        SmoothFunction(spectrum, m_filter->value());
 
     series->clear();
     for (unsigned int i = 1; i <= spectrum->size(); i++) {
@@ -771,14 +864,17 @@ void ThermogramWidget::ApplyCalibration()
         (m_peak_list)[i].integ_num -= m_const_offset->value();
     }
     UpdateTable();
-    emit Offset();
+    emit CalibrationChanged(m_calibration_heat->value());
 }
 
 void ThermogramWidget::UpdateBaseLine(const QString& str)
 {
+
     //if (str == m_base || m_block)
     //return;
     //m_thermogram->zoo
+
+    /*
     Vector coeff;
 
     if (str == "none") {
@@ -834,6 +930,7 @@ void ThermogramWidget::UpdateBaseLine(const QString& str)
     emit IntegrationChanged();
 
     m_base = str;
+    */
 }
 
 void ThermogramWidget::UpdateFit(const QString& str)
@@ -847,6 +944,25 @@ void ThermogramWidget::UpdateFit(const QString& str)
 
 void ThermogramWidget::FitBaseLine()
 {
+    Waiter wait;
+
+    PeakPick::spectrum* spectrum = new PeakPick::spectrum(m_spec);
+    PeakPick::BaseLine baseline(spectrum);
+    baseline.setBaseLineRange(PeakPick::BaseLine::BLR::PeakWise);
+    baseline.setPolynomFit(PeakPick::BaseLine::Polynom::Slow);
+    baseline.setNoCoeffs(2);
+    baseline.setPeaks(&m_peak_list);
+    m_baseline = baseline.Fit();
+
+    if (m_baseline.baselines.size() == 1) {
+        m_initial_baseline = m_baseline.baselines[0];
+        m_coeffs->setValue(m_initial_baseline.size());
+    }
+    delete spectrum;
+
+    return;
+
+    /*
     if (m_baseline_type->currentText() != "polynomial")
         return;
     Waiter wait;
@@ -887,7 +1003,7 @@ void ThermogramWidget::FitBaseLine()
 
     baseline.setInitialBaseLine(m_initial_baseline);
 
-    /* QThread *thread = QThread::create([this, baseline]{ */ m_baseline = baseline.Fit(); /*});*/
+    m_baseline = baseline.Fit(); 
 
     if (m_baseline.baselines.size() == 1) {
         m_initial_baseline = m_baseline.baselines[0];
@@ -895,6 +1011,7 @@ void ThermogramWidget::FitBaseLine()
     }
 
     delete spectrum;
+    */
     //Update();
     //emit IntegrationChanged();
 }
@@ -977,6 +1094,7 @@ void ThermogramWidget::PeakDoubleClicked(int peak)
     m_thermogram->setYRange(ymin, ymax);
 
     m_thermogram->setSelectBox(QPointF(xmin, ymax), QPointF(xmax, ymin));
+    setGuideText(QString("You are now in <i>Peak Integration mode</i>. Click [ESC] to leave to mode. Sometimes you might have to activate the chart widget by <b>clicking</b> with the <b>left mouse button</b> right before [ESC]. The peak integration range can be reduced by clicking the <b>right mouse button</b> within the black separation lines - or reset to the whole peak <b>clicking</b> with the <b>right mouse button</b> out of the area between the separation borders. Use [LEFT] or [RIGHT] arrow on your keyboard to navigate through the peaks or <b>double-click</b> on a different peak in the table. Zooming with the <b>mouse wheel</b> is possible."));
     //m_thermogram->setFocus(Qt::MouseFocusReason);
 }
 
@@ -994,17 +1112,17 @@ void ThermogramWidget::PeakChanged(int row, int column, int value)
 void ThermogramWidget::setFit(const QJsonObject& fit)
 {
     m_block = true;
-    m_baseline_type->setCurrentText(fit["baseline"].toString());
-    m_fit_type->setCurrentText(fit["baseline_fit"].toString());
-    m_coeffs->setValue(fit["coeffs"].toInt());
+    // m_baseline_type->setCurrentText(fit["baseline"].toString());
+    // m_fit_type->setCurrentText(fit["baseline_fit"].toString());
+    // m_coeffs->setValue(fit["coeffs"].toInt());
     m_const_offset->setValue(fit["constants"].toDouble());
-    m_stdev->setText(fit["stddev"].toString());
-    m_mult->setText(fit["multiplier"].toString());
+    // m_stdev->setText(fit["stddev"].toString());
+    // m_mult->setText(fit["multiplier"].toString());
     m_integration_range_threshold->setValue(fit["integration_range_threshold"].toDouble());
     m_integration_range->setCurrentText(fit["integration_range"].toString());
     m_initial_threshold = fit["integration_range"].toDouble();
 
-    if (fit.contains("smooth")) {
+    /*  if (fit.contains("smooth")) {
         m_smooth->setChecked(true);
         m_filter->setValue(fit["SV"].toInt());
     }
@@ -1017,6 +1135,7 @@ void ThermogramWidget::setFit(const QJsonObject& fit)
         m_poly_slow->setChecked(fit["slow"].toBool());
     } else
         m_poly_slow->setChecked(false);
+    */
     m_peaks_start->setValue(fit["start_time"].toDouble());
     m_peaks_time->setValue(fit["peak_time"].toDouble());
     m_peaks_end->setValue(fit["end_time"].toDouble());
@@ -1043,10 +1162,9 @@ void ThermogramWidget::setFit(const QJsonObject& fit)
 QJsonObject ThermogramWidget::Fit() const
 {
     QJsonObject fit;
-    fit["baseline"] = m_baseline_type->currentText();
+    /* fit["baseline"] = m_baseline_type->currentText();
     fit["baseline_fit"] = m_fit_type->currentText();
     fit["coeffs"] = m_coeffs->value();
-    fit["constants"] = m_const_offset->value();
     fit["stddev"] = m_stdev->text();
     fit["multiplier"] = m_mult->text();
     if (m_smooth->isChecked()) {
@@ -1054,7 +1172,9 @@ QJsonObject ThermogramWidget::Fit() const
         fit["SV"] = m_filter->value();
     }
     fit["peakwise"] = m_peak_wise->isChecked();
-    fit["slow"] = m_poly_slow->isChecked();
+    fit["slow"] = m_poly_slow->isChecked(); */
+
+    fit["constants"] = m_const_offset->value();
     fit["frequency"] = m_frequency;
     fit["start_time"] = m_peaks_start->value();
     fit["end_time"] = m_peaks_end->value();
@@ -1108,8 +1228,8 @@ void ThermogramWidget::CreateSeries()
 void ThermogramWidget::UpdatePeaks()
 {
     qint64 t0 = QDateTime::currentMSecsSinceEpoch();
-    qreal offset;
-    int off = 0;
+    qreal offset = 0;
+    int off = 1;
     //double start = m_peaks_start->value();
     int rules_size = m_peak_rule_list->rowCount();
     double start = m_peak_rule_list->item(0, 0)->data(Qt::UserRole).toDouble();
@@ -1121,7 +1241,11 @@ void ThermogramWidget::UpdatePeaks()
         QTableWidgetItem* item = (m_peak_rule_list->item(j, 0));
         double index_start = m_spec.XtoIndex(item->data(Qt::DisplayRole).toDouble()); //m_spec.XtoIndex(m_peak_rule_list->item(j, 0)->data(Qt::UserRole).toDouble());
         item = m_peak_rule_list->item(j, 1);
-        double timestep = m_spec.XtoIndex(item->data(Qt::DisplayRole).toDouble()); //m_spec.XtoIndex(m_peak_rule_list->item(j, 1)->data(Qt::UserRole).toDouble());
+        double timestep = m_spec.XtoIndex(item->data(Qt::DisplayRole).toDouble());
+        if (timestep <= 0) {
+            m_guide_label->setText(QString("<font color='red'>Sorry, but Peak rule %1 contains a zero as peak time. That means, if I did not stop right here, you would be waiting for Godot</font>").arg(j + 1));
+            return;
+        } //m_spec.XtoIndex(m_peak_rule_list->item(j, 1)->data(Qt::UserRole).toDouble());
         double index_end;
 
         if (j == rules_size - 1)
@@ -1150,6 +1274,8 @@ void ThermogramWidget::UpdatePeaks()
             Update();
 
         emit PeaksChanged();
+        ResetGuideLabel();
+
         // emit IntegrationChanged();
 }
 
@@ -1170,29 +1296,35 @@ void ThermogramWidget::AddRectanglePeak(const QPointF& point1, const QPointF& po
     //qDebug() << m_peak_list[m_current_peak].start << m_peak_list[m_current_peak].int_start << m_peak_list[m_current_peak].int_end << m_peak_list[m_current_peak].end;
     FitBaseLine();
     Update();
+    ResetGuideLabel();
 }
 
 void ThermogramWidget::PointDoubleClicked(const QPointF& point)
 {
     QApplication::restoreOverrideCursor();
 
-    disconnect(m_thermogram, &ChartView::PointDoubleClicked, this, &ThermogramWidget::PointDoubleClicked);
-    if (m_get_time_from_thermogram == 1) {
-        m_peaks_start->setValue(point.x());
-        m_peak_rule_list->item(m_current_peaks_rule, 0)->setData(Qt::DisplayRole, point.x());
-    } else if (m_get_time_from_thermogram == 2)
-        m_peaks_end->setValue(point.x());
-    Divide2Peaks();
-    m_get_time_from_thermogram = 0;
-}
+    qreal x = point.x();
 
-void ThermogramWidget::Divide2Peaks()
-{
-    return;
-    if (m_peaks_start->value() < m_peaks_end->value()) {
-        std::vector<PeakPick::Peak> peaks = PeakPick::Divide2Peaks(&m_spec, m_peaks_start->value(), m_peak_count->value(), m_peaks_end->value());
-        setPeakList(peaks);
+    if (m_get_time_from_thermogram == 1) {
+        m_peaks_start->setValue(x);
+        m_peak_rule_list->item(m_current_peaks_rule, 0)->setData(Qt::DisplayRole, x);
+        m_get_peaks_start->setText("Click to Select");
+    } else if (m_get_time_from_thermogram == 2) {
+        m_get_peaks_end->setText("Click to Select");
+        m_peaks_end->setValue(x);
+    } else {
+        for (int i = 0; i < m_table->rowCount(); ++i) {
+            double start = m_table->item(i, 1)->data(Qt::DisplayRole).toDouble();
+            double end = m_table->item(i, 2)->data(Qt::DisplayRole).toDouble();
+
+            if (start < x && x < end) {
+                m_table->setCurrentCell(i, 0, QItemSelectionModel::Rows | QItemSelectionModel::SelectCurrent);
+                break;
+            }
+        }
     }
+    ResetGuideLabel();
+    m_get_time_from_thermogram = 0;
 }
 
 void ThermogramWidget::CalibrateSystem()
@@ -1225,7 +1357,7 @@ void ThermogramWidget::CalibrateSystem()
     m_calibration_peak.integ_num = PeakPick::IntegrateNumerical(spectrum, m_calibration_peak.start, m_calibration_peak.end, baseline_result.baselines[0]);
     // qDebug() << m_calibration_peak.integ_num<< m_calibration_peak.start<< m_calibration_peak.end;
     delete spectrum;
-    m_calibration_label->setText(tr("Calibration (%1)").arg(m_calibration_peak.integ_num));
+    m_calibration_label->setText(tr("<h4>Calibration (%1)</h4>").arg(m_calibration_peak.integ_num));
 }
 
 bool ThermogramWidget::CutAllLimits()
@@ -1294,6 +1426,8 @@ bool ThermogramWidget::CutAllLimits()
 
     delete spectrum;
     Update();
+    setGuideText(QString("Adaption of the baseline took %1 cycles").arg(m_last_iteration_max));
+
     return true;
 }
 
@@ -1329,4 +1463,20 @@ void ThermogramWidget::scaleDown()
         m_thermogram->setYRange(min + next, max);
     else
         m_thermogram->setYRange(min - (ratio)*next, max + (1 - ratio) * next);
+}
+
+void ThermogramWidget::setGuideText(const QString& str)
+{
+    if (qApp->instance()->property("thermogram_guidelines").toBool())
+        m_guide_label->setText(str);
+    else
+        m_guide_label->clear();
+}
+
+void ThermogramWidget::ResetGuideLabel()
+{
+    if (qApp->instance()->property("thermogram_guidelines").toBool())
+        m_guide_label->setText("You are in standard mode. In addition to the SupraFit chart interaction, the <b>mouse wheel</b> can be used to zoom in and out, where the ratio of zooming depends on the actual positon of the curser in the chart. See the SupraFit Handbook for details. And <b>double click</b> in the chart selects the peak in the table.\nRegarding the units in this tabulator: They are all in the units observed by your calorimeter. If they are originally for example in cal, please mind the scaling factor to convert them into J. If they are already in J, since you are using a calibration peak, set the scaling factor to 1 (this may have been done automatically, but always check).");
+    else
+        m_guide_label->clear();
 }